// src/types/actionTypes.ts
export enum ACTION_TYPES {
  FETCH_START = 'FETCH_START',
  FETCH_SUCCESS = 'FETCH_SUCCESS',
  FETCH_ERROR = 'FETCH_ERROR',
  EDIT_START = 'EDIT_START',
  CREATE_SUCCESS = 'CREATE_SUCCESS',
  EDIT_ERROR = 'EDIT_ERROR',
  REMOVE_SUCCESS = 'REMOVE_SUCCESS',
  EDIT_SUCCESS = 'EDIT_SUCCESS',
  CREATE_ITEM_SUCCESS = 'CREATE_ITEM_SUCCESS',
  TXN_FETCH_START = 'TXN_FETCH_START',
  TXN_FETCH_SUCCESS = 'TXN_FETCH_SUCCESS',
  TXN_FETCH_ERROR = 'TXN_FETCH_ERROR',
  TXN_EDIT_START = 'TXN_EDIT_START',
  TXN_CREATE_SUCCESS = 'TXN_CREATE_SUCCESS',
  TXN_EDIT_ERROR = 'TXN_EDIT_ERROR',
  TXN_REMOVE_SUCCESS = 'TXN_REMOVE_SUCCESS',
  TXN_EDIT_SUCCESS = 'TXN_EDIT_SUCCESS',
}